backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
textblob==0.18.0
vaderSentiment==3.3.2
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
httpx==0.27.2
python-dotenv==1.0.1
